feat!: replace AccessTokenFactory with AbstractAccessTokenRepository
Introducing this change with the intent of making the proxying token storage
agnostic. The presumption is that user identification is included, potentially
as an attribute appended by middleware, in the request, and that there is
the possiblilty that some additional identification may need to be sent with
responses. AccessTokenCookie is provided as a default implementaton for
implementations of ProviderInterface to make use of, storing access tokens
in browser cookies.

BREAKING: implementations of ProviderInterface must implement
setAccessTokenReposiitory() and getAccessTokenRepository(). A default trait
`Defaults\AccessTokenRepository` is provided that uses an AccessTokenCookie